requirements.txt
psycopg2-binary==2.9.9
networkx==3.1
matplotlib==3.7.1
sqlparse==0.4.4
pytest==7.4.3
PyQt5==5.15.9
flask==3.0.0